LLM_Server/requirements.txt
# Core Python packages
python-dotenv>=0.19.0

# Database
pymongo>=4.0.0
certifi>=2021.10.8

# LLM and AI
openai>=1.0.0

# Web frameworks and HTTP
flask>=2.0.0
requests>=2.25.0
requests-cache>=0.9.0
retry-requests>=2.0.0

# Weather and Geospatial
openmeteo-requests>=1.0.0
geopy>=2.2.0

# Data Processing
pandas>=1.3.0
numpy>=1.21.0

# Logging (already in Python standard library)
# logging - built-in

# Other standard library modules used (no installation needed):
# json, re, os, sys, datetime, typing, time

# Development and Testing
# No specific testing frameworks detected, using built-in unittest/logging

# Optional: Additional packages that might be useful for production
urllib3>=1.26.0
httpx>=0.24.0  # Used by openai library
